requirements.txt
Flask==3.0.3
click==8.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.1.3
MarkupSafe==2.1.1
python-dotenv==0.17.1
Werkzeug==2.2.3
openai
pymongo
flask_cors
requests
ngrok
twilio
python-dotenv